Move constants to constants.ts
