+++ packages/web3torrent/src/constants.ts
@@ -1,7 +1,27 @@
 import {Status, Torrent} from './types';
 import {ChannelState} from './clients/payment-channel-client';
-
 import {bigNumberify} from 'ethers/utils';
+
+export const WEI_PER_BYTE = bigNumberify(1); // cost per byte
+export const BUFFER_REFILL_RATE = bigNumberify(2e4); // number of bytes the leecher wishes to increase the buffer by
+// These variables control the amount of (micro)trust the leecher must invest in the seeder
+// As well as the overall performance hit of integrating payments into webtorrent.
+// A high BUFFER_REFILL_RATE increases the need for trust, but decreases the number of additional messages and therefore latency
+// It can also cause a payment to go above the leecher's balance / capabilities
+export const INITIAL_SEEDER_BALANCE = bigNumberify(0); // needs to be zero so that depositing works correctly (unidirectional payment channel)
+export const INITIAL_LEECHER_BALANCE = bigNumberify(BUFFER_REFILL_RATE.mul(100)); // e.g. gwei = 1e9 = nano-ETH
+
+// firebase setup
+export const HUB_ADDRESS = 'TODO';
+export const FIREBASE_PREFIX = 'web3t';
+export const fireBaseConfig = {
+  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
+  authDomain: `${process.env.REACT_APP_FIREBASE_PROJECT}.firebaseapp.com`,
+  databaseURL: `https://${process.env.REACT_APP_FIREBASE_PROJECT}.firebaseio.com`,
+  projectId: process.env.REACT_APP_FIREBASE_PROJECT,
+  storageBucket: '',
+  messagingSenderId: '913007764573'
+};
 
 export const defaultTrackers = [
   // 'udp://explodie.org:6969',
